// src/app/models/Grafic_req.ts
export class Grafic_req {
  idGrafic?: String;
  iar?: String;
  code_imb?: String;
  groupe_operation?: String;
  dateTraitement?: Date | null;
  statut_graphic?: String;
  traitement_effectue?: String;
  type_traitement?: String;
  active?: boolean;
  idacte?: String;
  affectation?: String;
  duree?: Number;
  commentaire?: String | null;
}